+++ src/config/index.ts
@@ -4,7 +4,8 @@
  * @license Apache 2.0
  */
 
+export { GlobalConfigManager } from './global-config';
+export { ProjectConfigManager } from './project-config';
+export { ConfigValidator } from './validation';
+export { ManifestValidator } from './manifest-validator';
 export * from './constants';
-export * from './global-config';
-export * from './project-config';
-export * from './validation';